Fixed what I believe was a bug in `FakeSockFile`.
It was returning padding spaces when an `amount` greater than the length
of the body was sent to `read`. This makes it impossible to find the end
of a response in the streaming case.

While all unit and functional tests pass after this change, I'm not sure
whether some use cases rely on the old padding behavior. If that's the
case, streaming should be special-cased- perhaps by adding a
`read_streaming` or similar method to `FakeSockFile`.