feat: leaner requirements, improved tox/travis support
This patch reorganizes the requirements to install only what's
necessary for using httpretty. This greatly reduces installation time
when being pulled in as a dependency for another project.

The dependencies needed for running the test suite were moved to
test-requirements.txt, and were commented to indicate what type of
testing dependency it it.

The tox file and the travis yaml have been updated to reflect the
changes in the requirements files.

Finally, .travis.yaml has leveled up. It now supports running the test
suite using pypy and python 3.3, with the added benefit that a build
will still be marked as successful even if py3.3 and pypy fail.