feat(setup): migrate to using pbr
This patch provides support for using pbr to manage the packaging
metadata for HTTPretty.

It removes the need for a custom requirements parsing function, and
unifies support for distutils and setuptools-based installation.

The primary obvious downside is that version management is now a
manual process. Rather than being determined by the httpretty.__init__
module, it must be set in the setup.cfg before cutting a release.

This merges the support for [nosetests] configuration given in the
current setup.cfg and updates the tox.ini accordingly.