Rename readme.rst
By default, setup.py will package README.rst.  If the package is built
on a case-insensitive filesystem, this would mean that readme.rst would
be distributed as README.rst, resulting in file read errors when
attempting to install the resulting package on a system with a case-
sensitive filesystem.  This corrects the issue by the simple expedient
of renaming the file.

Addresses #176.