Add priority argument to register_uri.
Matchers are stored in a dictionary and are checked in an unpredictable
order. This can be a problem when multiple matchers match the same
URI--for example, when using multiple regex matchers. This patch adds an
optional an optional priority argument to register_uri to resolve this
ambiguity. When matching a URI, matchers are sorted descending by
priority, and the first match is used.