Handle bugs in older urllib3 versions in one of the tests
Older urllib3 versions had a bug where they lower-cased header
names (in response header dicts). That makes one of our tests
fail with older urllib3, because the test expects a 'Server'
header. As this isn't our fault at all, just have the test cope
with it by checking if the header dict has a 'server' key and
replacing it with a 'Server' key with the same value.

urllib3 1.10 also had a bug when you called dict() on its
HTTPHeaderDict class; it would turn this:

{'headername': 'value'}

Into this:

{'headername': ['headername', 'value']}

That was fixed in 1.11, but RHEL 6 still has 1.10, so let's
work with that by doing dict(headerdict.items()) instead of
just dict(headerdict) (when we're recording the calls).